Move ObjectContactsView to netbox.views.generic.feature_views
