+++ netbox/tenancy/views.py
@@ -1,34 +1,11 @@
 from django.contrib.contenttypes.models import ContentType
 from django.shortcuts import get_object_or_404
-from django.utils.translation import gettext_lazy as _
 
 from netbox.views import generic
 from utilities.query import count_related
-from utilities.views import GetRelatedModelsMixin, ViewTab, register_model_view
+from utilities.views import GetRelatedModelsMixin, register_model_view
 from . import filtersets, forms, tables
 from .models import *
-
-
-class ObjectContactsView(generic.ObjectChildrenView):
-    child_model = ContactAssignment
-    table = tables.ContactAssignmentTable
-    filterset = filtersets.ContactAssignmentFilterSet
-    filterset_form = forms.ContactAssignmentFilterForm
-    template_name = 'tenancy/object_contacts.html'
-    tab = ViewTab(
-        label=_('Contacts'),
-        badge=lambda obj: obj.get_contacts().count(),
-        permission='tenancy.view_contactassignment',
-        weight=5000
-    )
-
-    def dispatch(self, request, *args, **kwargs):
-        model = kwargs.pop('model')
-        self.queryset = model.objects.all()
-        return super().dispatch(request, *args, **kwargs)
-
-    def get_children(self, request, parent):
-        return parent.get_contacts().restrict(request.user, 'view').order_by('priority', 'contact', 'role')
 
 
 #
